requirements.txt
langgraph>=0.2.50
langchain-openai>=0.2.0
langchain-core>=0.3.0
arize-phoenix>=5.0.0
arize-phoenix-otel>=0.6.0
openinference-instrumentation-langchain>=0.1.30
python-dotenv>=1.0.0
